Refactor Foods items components from the project
