--- src/components/Meals/Meals.js
@@ -1,13 +1,0 @@
-import React from 'react';
-
-import AvailableMeals from './AvailableMeals';
-
-const Meals = () => {
-  return (
-    <React.Fragment>
-      <AvailableMeals />
-    </React.Fragment>
-  );
-};
-
-export default Meals;
